Added Reaction to the End of Dialogues
Previously, when our agents would finish talking with someone, they wouldn't react to the conversation. If the conversation would suggest going to a specific location or changing plans, the agent would ignore it entirely.

Now, in the end_dialogue() function, there is a quick prompt to see if there was anything important in the full dialogue history that this agent should react to. If the response is "React", we run _pick_location() to see where to go considering the new action to take and also _generate_plan() to make sure they account for this new reaction.

The rest of the end_dialogue() function remains the same.