Added Item Interactions for Agnets
When agents enter the interaction zone of the item they want to interact with, they prompt chatgpt to come up with the new status of the item. Once they get the response, the item's action will be set to the new status and the item's interactable will be set to the agent that wants to interact with it.

We use this interaction to prevent multiple agents from interacting with the same item. Inside the get_sub_locations() function of the game manager, we make sure that if an item is currently being interacted with, it will not be presented as an option of a place to go to when the agents ask for it.