Fixed Bug: Agents no longer interact with the same items
Previously, when an agent wanted to pick a location or item to go to, it would consider all nodes in the layout. This meant that two agents could decide to interact with the same item, leading to the agents standing at the same location and updating the status of the item on their own.

Now, the agents would navigate to their destination, but once they see their target item is already interacting with another agent/player, they stop in their tracks. This means they "discover" their target is unavailable upon arrival.

To implement this, I have made the following changes:

- In the Brain's _on_interaction_zone_area_entered() function, if the item_entity already had an interactable, remove the target destination

- When the target item is reached in _on_interaction_zone_area_entered(), the item_entity's interactable is set to the agent's entity. When the agent leaves an item in _on_interaction_zone_area_exited(), the item_entity's interactable is reset to null

- When the player interacts with an item in Player.gd, the item's interactable is set to the player's entity. When the player leaves an item, the game manager resets the item's interactable to null in close_item_panel()

Also I renamed "Item Panel.gd" to "Item_Panel.gd" to remain consistent with the naming convention