Updated Luna's Color Palette + Refactored Animation Code
Luna's color palette clearly stood out amongst the environment and the rest of the agents, so I tried to tone her colors down a bit and make it more in-theme.

Also, I removed the need to have a dedicated icon image for each agent. Now, the inside the Animation_Creator.gd, the agent's icon texture is picked out from the assigned character texture file, so it'll do it automatically.

Also, the agents names in the Node hierarchy is changed to "Agent 1", "Agent 2", and so on, since their actual names may change.