Repainted the Physics Layer for the TileSet
Some of the tiles in the TIleSet we used had some empty space that still was assigned a physics collider, meaning the player and agents couldn't walk through certain tiles that seemed to be mostly empty space. So I repainted the physics layer of all the tiles we currently have in our environment, as well as a few other ones.