Inventory system added
Added an inventory system that contains our resources, till now the only resource is a pistol which i added to test out and we can easily add other items by including them in the resource, inventory is a little glitched the size and positioning is a bit weird, but hopefully would be an easy fix. Inventory can be opened or closed by pressing "p".