Added Continuous Tracking of Nearby Entity
Previously, the agents would observe the world around them only at one instant when their brain is triggered to activate. This means the agents would be blind for 10 seconds and then only see the world for a split second, which isn't realistic.

Now, the agents keep track of all new observations they have come across since their last trigger, including different observations on the same entity. For example, if Amelia goes from the living room to the meeting room, and interacts with a lamp, 3 new observations will be recorded: Amelia is in the living room. Amelia is in the meeting room. Amelia is in the meeting room and is interacting with the lamp.

The specific changes are as follows:

- The Entity class now stores a reference to the node it is attached to. This is important because as we store multiple different versions of the same entity in the new_observations list, we will eventually need a reference to the Node that is related to the given entity.

- Updated the agent, item, and player entity creation to include a reference to themselves (including the "self" keyword in the parameters)

- The Entity class now has a function called copy(), which returns a new object with the exact same values as the original. Used to create multiple observations of the same entity.

- The Entity class now has a function called matches(), which returns true if the given entity matches the original one. Used to compare between existing and new observations.

- The _observe() function is called once per frame in the _physics_process() function of Brain.gd, allowing for continuous observation

- The _observe() function now checks to see if a new observation matches any of the existing ones using the match() function, and only when we have found a unique observation will we add it to the list

- The new_observations array now stores Entity objects instead of Node objects