Inventory system added
We have an inventory system that we could actually save in agents and the player itself, as the inventory is a resource, i also added a pistol and a coffee mug as resources to the player inventory, we can have as many items as possible we just need to add them as "InvItem" resources which we only need to provide a name and a texture for them. we can use "p" to open and close the inventory, i've hard coded two pistols and a mug to the inventory just to see how it looks, we can't pick them up and add them automatically yet, i saw that the pistol is somewhat a good fit into the inventory box but the coffee mug looks a bit small but should be easily fixable.