Implemented Changing Destination from Afar when It's Occupied
Previously, an item should have been inside an agent's interaction zone for the agent to consider if the item was taken or not by some other agent/player. This lead to a non-natural way of updating the path.

Now, the agent will use the continuous observation as the way it detects if the target destination is taken or not. If the entity that is in the path of the ray is the destination, then we check if it's taken or not. If it is, we pick a similar entity in the same room that's unoccupied.

I've implemented the following changes:

- The agent's navigation is now Edge Centered, to prevent the agents from getting stuck on walls

- The _observe() function is now called once every 5 frames instead of every frame, this is to reduce stress on the computer's resources

- The code to change destination is moved from the _on_interaction_zone_body_entered() function to the _observe() function