Created Setup Scene, Panel, and Grid Container
When we start up our game, we need to choose our player and generate the personalities/history of all our agents. That will be done in the new "Setup" scene.

In this scene, we have the following:

- A Game Manager which handles all the logic of the panel interaction.

- A Panel that directs the player to do different tasks and contains all the characters that we can present to the player.

- A Grid Container inside a Scroll Container, which allows our characters to be displayed in an ordered fashion and can be scaled up to any number of characters (not just 5).

- The Grid Container will load up all character png files that are now renamed to "Agent_0.png", "Agent_1.png", and so on and add them to the grid. That way, it dynamically adds the characters.

For now, there is no user input detected. But our grid does successfully load up the agent files and presents them properly.