Fixed Loading the Gender of Agent Image Sheets
Previously, the gender of each agent was determined by their position in the character_nodes array. If they were in the first half, they were male, and if they were in the second half, they were female. This already had an issue with the midway point of the array, which could be either a male or a female. Also, this file naming convention meant that we couldn't easily insert new agents into the folder without having to rename a bunch of other files and figuring out what the new index should be.

To solve this, I came across the DirAccess class which provides functionality regarding reading and writing to a directory. With this, I propose a new naming convention for our agent image sheets: "Agent_[M/F]_[other info].png". This way, the file will store the gender of the agent itself, and it doesn't care about how many other agents exist in the directory. The [other info] field can be names, numbers, whatever. It won't be ready by the application.

I have made the following changes:

- Added a "character_genders" array to the Setup_Scene.gd script which stores whether or not each new character we read is a male (true) or female (false). This is used later when we want to generate a random name to quickly find the gender.

- All character image sheets now conform to the new format, such as "Agent_M_Alex.png" and "Agent_F_Luna.png". Note that the agent name is only used for our organization and clarity, they have no affect on the game.