Created Game_Manager Inheritance Structure + Refactored Concurrency Code
-- Part 1 --

Previously, the Setup_Scene.gd script inherited from the Game_Manager.gd script so that it can also make API calls when it needs to. This worked; however, it meant all the game logic and node variables used by the Game_Manager.gd (like item panels, player, and so on) would be accessible by the Setup_Scene.gd script, which caused a few errors and wasn't good code design.

Now, I have turned Game_Manager.gd into a base class that all scene scripts will inherit from, and it will only contain code that's relevant for all scenes (like API calls). To do this, the GameManager node in the Game Scene now has a new script called Game_Scene.gd that inherits from Game_Manager.gd. All the code related to the main game functionality is in there.

-- Part 2 --

Previously, the way we made concurrent calls to the APIs in the Brian.gd script was using the "callback_signal" to register every time a request was completed, and we'd await for the _collect_responses() function to go through all submitted requests before proceeding in the program. This had a few problems. First, when we create a new Memory object, the callback_signal will get triggered every time, even if it's not necessary to wait for the memory to be finished. This means it could interfere with other processes when they're waiting for the responses to be completed. Second, this concurrent behavior is very useful, but it currently only exists in the Brain.gd script. We need this in the Setup_Scene.gd script too, so ideally this should be a separate class.

Now, the concurrency logic is handled by a dedicated class called ConcurrencyHandler in the Concurrency_Handler.gd script. It has the same logic as the previous implementation, but now it uses 3 special functions:

- response_complete(): when we reach a point in the code where the request has been completed, we call this function to update the handler

- wait_for_responses(): this is the renamed version of "_collect_responses()". If called with the await keyword, the program will stop until all requests have been completed before continuing.

- pop_responses_as_string(): combines all the responses into one string seperated by a given delimiter and clears the responses array

Note how the _add_memory() function can now use a second argument "is_concurrent" to only trigger the concurrency handler for memories that matter. By default, adding memories doesn't stop the program. But initially, when we want to generate the agent summar in _ready(), we need to make sure the existing memories are added before doing so.