Added Data Transfer Between Scenes
Previously, when you transitioned form the Setup Scene to the Game Scene, no data that was generated in the Setup Scene was carried over.

Now, I have added an AutoLoad item called DataTransfer, based on the new Data_Transfer.gd script, which will store all that we need to transfer between scenes. I've also made it so that the game scene dynamically adds new agents to the world based off of the files that were loaded up in the Setup Scene, which means we don't have to manually add new agents, it will do it automatically.

I have implemented the following changes:

- Right now, all we need is information about the player and information about all the other agents, which are what the two variables represent in the Data_Transfer.gd script. The functions in that script help us access the variables and set their values.

- We set the values of DataTransfer when we click the start button. I've organized the values in multilpe Dictionary objects for easier readability and management. Then, in the _ready() function of the Game_Scene.gd script, we read the DataTransfer values and assign them to our agents/players.

- Since new agents will be generated on the fly, we need a way of picking where they spawn in the world. Fortunately, we already have standing zones, so I have added two new functions to Game_Scene.gd: get_all_standing_zones() and _add_standing_zones_to_list(). These work together to create a list of all standing zones in the world, and we randomly pick one for the agent to spawn at. Note that overlaps my happen between two different agents.

Note that if you run the game normally (i.e. in the Game scene), then the player will be Alex and the only agent will be Luna. This is set using the default values of the DataTransfer class so that starting up the Game_Scene.gd doesn't crash. This is purely for debugging purposes if we want to access the Game scene without having to go through the setup. If you want to get all agents, start from the Setup_Scene.