Fixed Character Sheet Loading Issue during Setup
Previously, the Setup_Scene.gd file would load all files in the "Characters" folder except for ones with ".import" in them. This worked, but now a new file type that ends with ".png~" has caused the loading to crash since it's not a valid character sheet.

Now, to make the loading more robust, I have made it so that the code will skip anything that doesn't end with ".png". This fixed the issue.