Added More Test Agents for the Game Scene
Previously, we only had Luna that would exist by default in the Game Scene if we didn't go through the setup process. This was not enough to test agent interactions and perform practical testing.

Now, we've included all the agents in the game scene, so we don't need to go through the setup process and can perform testing as we see fit. There isn't any history generated for each character; if we need history, we must go through the setup process.