Attached a button to the slot scene
Each slot contains a button of its own, and whenever a slot is not empty the button gets activated and ready to be clicked.

when a button is clicked it'll return the "method_of_killing" variable that exists inside our resource script, this can be initialized whenever we make a new resource that we want to add to our main scene.