Decorated the Outside of the Mansion with Nature, Walls, Roof, and a Path
Previously, the outside of the mansion would just show a flat green color since that was the default background of the project.

Now, I have added a new tile sheet called "RPG_basics" that has nature tiles and some building tiles. Using this tile sheet, the surroundings of the mansion are now lushous and more interesting, and there are proper walls and roof to the mansion.