Player Movement is Now Smoother
Previously, the player could move around in only 8 possible directions, leading to robotic movement and varying speeds. Also, the movement had no easing in or out, so it was quite jarring.

Now, the Player.gd script contains an "acceleration" vector that determines what direction the player should be heading towards, which is determined based on the movement buttons pressed. This acceleration is added to the velocity so that over time the speed increases and it looks like it's easing into the movement. When the player doesn't press any movement buttons, the velocity slows down towards zero.