"Interaction" UI Now Indicates which Entity to Interact with
Previously, the "I" symbol on top of the player would only indicate if an entity has entered the player's interaction zone. This meant that if there were multiple nearby entities, there would be no indicator of which one we could interact with since the label never changes. Also, relying on only the entering and exiting of the interaction zone's collision shape meant that sometimes we won't register an entity that's already in the interaction zone without having to leave and enter again.

Now, the nearby entity detection is done continuously in the _physics_process() function of Player.gd, which means we no longer use the signals and functions of our interaction zone Area2D to find entities. Also, to indicate which entity we can interact with, the "I" label will point towards the entity and follow it around as you move, clearly showing what you can interact with.