Game Manager Now Keeps Stats in Every Stage
Previously, the statistics of how many times we called the APIs would only be shown in the game scene, and updated only once every second. Also, the UI for that information was always visible.

Now, the GameManager has it's own scene, where it contains the stats label itself. This way, we can display stats and info in any stage, not just the game scene. The panel can be toggled on and off with the Ctrl+S key combination.

Also, the stats are now updated every time the API is called and once per second.