Added Killing Indicator for Player Popup
Previously, the popup UI for the player would only show "I" to indicate that the player can interact with that entity. It would not show the "K" option for killing agents.

Now, the label updates to show only "I" for items and "I , K" for agents. Also, dead agents aren't considered as interactable entities anymore.