Player can Now Only Kill the Target Agent
Previously, the player could kill any agent it came close too. This would require some effort to handle what happens when the player kills multiple agents, which isn't a priority right now.

Now, the Brain.gd script has an "is_target" boolean variable that is true if the agent is the person the player has to kill. To set this, the setup scene will assign the value based on the random agent it picked as the target.

Note: If we start from the Game Scene directly, Bob is the target every time.