Corrected Wall Physics Masks for Better Sliding
Previously, the markers of the physics masks of the wall tiles were sometimes offset by small amounts, making the game think the walls weren't straight and instead was slanted. This lead to staggered sliding along the walls when multiple tiles were placed next to each other.

Now, the physics mask for the wall tiles have been repainted to be discrete accurate rectangular blocks which now has fixed the staggering.