Cached what an Agent Knows about Dialogue Partner
Previously, the information about what an agent previously knows about their dialogue partner was generated each time a statement was exchanged in the dialogue. This was unnecessary since no knew memories would be added that could change the result.

Now, the variable dialogue_context stores the generated information about the dialogue partner in Brain.gd and it gets referenced every time the agent wants to come up with what to say.