Added Smoother Transition between Day/Night Colors
Previously, the day/night colors were picked based on the integer value of the current hour. This worked, but because a lot can change in an hour, you would see a drastic change in the color between, say, 8:59 am and 9:00 am.

Now, the color is picked based on a floating point value that combines the current hour and minute to generate a value between 0.0 - 24.0, leading to smoother transitions.