Added a "Meeting" UI
once the meeting UI is activated the game will show a panel with the text "Dead body found", and after 6 seconds that panel will close and another panel will be shown, on the left side of the panel there'll be 5 names which are the agents and the player themselves to be kicked out, and we have a skip vote option as well. on the right side will be our chat where the player and agents can talk to each other before proceeding to vote someone out.