Slightly Improved Calculating Cosine Similarity
Previously, the bottleneck of memory retrieval was the _calculate_cosing_similarity() function in Brain.gd, because even for our small embedding vectors, we would have 1536 dimensions to compute a lot of times.

To improve this, we removed the square root inside the function as this is a computationally expensive function. This seems to have improved the performance a bit.

Note: performance lag is still clearly visible, and would need to be readdressed because it can get very extreme once these agents get a lot of memories. For now, this'll do.