Added Speech Bubble to Group Meeting
Now we have a Speech Bubble scene that will display what each character says in the group meeting. This scene is preloaded in Brain.gd and Player.gd and can be instantiated anytime a new piece of text enters the convo.

To create a new speech bubble inside Brain.gd or Player.gd, call the _add_speech_bubble() function that takes in the text inside the bubble and whether or not the speech bubble is placed to the left of the screen and the top.

Once a new bubble is created, it will grow using a bounce animation and will exist for the lifetime_frames number of frames. Then, when the lifetime is over, it will shrink using a bounce animation and get removed from the scene.

Also, the group discussion panel is reworked to be more clean and useful. Right now only the chat panel has been updated, but the vote panel will also be updated soon.