Player's Speech Bubble Appears when Pressing ENTER
Previously, the UI for the chat panel had no functionality, meaning the player couldn't actually "talk". Also, the player would have to press the "talk" button in order for the text to be sent.

Now, there is no "talk" button, and by simply pressing the ENTER key, the entered text is shared with the agents as a new group message.

I have made the following changes:

- Player.gd and Brain.gd no longer create speech bubbles themselves. That function and resource is all handled by the Game_Scene.gd script now.

- Every time the text in the TextEdit of the chat panel is changed, the _on_text_edit_text_changed() function inside Meeting.gd is called. If the text contains a new line character, that means the player pressed the ENTER key and so it adds the contents of the TextEdit as a new group message.

- Game_Scene.gd now has a new function called add_group_message() that takes in the new message and the character who said that message. It currently creates a new speech bubble at the appropriate location.

- The Game_Scene.gd script has a new disable_all_UI() function that will turn off all UI elements, which can be useful when switching between modes and scenes.