Speech Bubble Duration now In Seconds
Previously, the speech bubble would remain on for a certain number of frames. This isn't an ideal implementation because it depends on the frame rate, which could change based on the device.

Now, the variable uses seconds to determine how long the speech bubble can be visible for. It is currently set to 5 seconds.