Added Bubble Timer + Agents Initiate Group Convo
Previously, the speech bubble would have a fixed lifetime of 5 seconds regardless of the length of text and it would not give an indication of when the bubble would close. Also, the scroll functionality of the bubbles seemed to not work.

Now, not only does the agent who called the meeting start the conversation, but the speech bubbles have a lifetime that depends on the length of the text and the average 250 words-per-min reading speed. Also, there's a progress bar to show how long the speech bubble has left.

Also, the scrolling issue was caused by a bunch of panels from the meeting scene being overlayed on the bubbles and taking priority over user input. Now, all panels are removed. These will be refactored lator.