Cleaned Up Voting Structure
Previously, the title of the voting panel would scroll with the buttons. Also, since the title was a part of the scroll container, the agent buttons would be too wide as it would have to match the width of the title.

Now, the title is separated from the scroll container and the buttons are centered and the right size.