Displaying Voting Results + Eliminating Characters
Now, the results from the voting process are tallied up and displayed to the player, including the number of skips. If a character got the highest number of votes, that character will be "arrested for investigation" and be removed from the game. If more than one character got the highest vote or the highest vote was "[Skip]", no one would be eliminated in that round. Depending on whether the player is eliminated or the player survives to the end, we print "LOSE GAME" and "WIN GAME" to mark the completion of the game.

Quick notes:

- The Game_Scene.gd script automatically removes all speech bubbles once the voting process starts

- If the player kills an agent and there is only 1 other alive agent left, the player wins