Fixed Bug where Agents Receive Tons of Discussions
Previously, the group discussion could get out of hand because when one agent would share information, all other agents would want to share information and this would repeat and grow very rapidly.

Now, if an agent hasn't finished generating the context of a previous discussion it heard, it will ignore a new discussion that it hears. This keeps the number of discussions generated in check.