Agents Now See Players Performing a Kill
Previously, the player's action was not changed when they killed an agent. Meaning it would look like the agent mysteriously died while the player was standing idle near them.

Now, the player's action is updated to reflect how they killed the agent. It is then reset to idle after 1 second.