"Enter Name" in Setup is Automatically In-Foucs
Previously, the player would have to navigate over with their mouse to the line edit element and click on it to enter their name, but swapping between mouse and keyboard can get annoying.

Now, the enter name element is automatically put in focus when the player selects their character, and pressing ENTER won't do anything if there is no name entered.